添加了删除权限限制

+++ delBlog.jsp
@@ -11,8 +11,17 @@ try {
 	Class.forName("com.mysql.jdbc.Driver");
 	String connectionUrl = "jdbc:mysql://172.18.187.253:3306/boke18329015?useUnicode=true&characterEncoding=UTF-8";
 	conn = DriverManager.getConnection(connectionUrl, "user", "123");
-	conn.setAutoCommit(false);
 	Statement stmt = conn.createStatement();
+	ResultSet rs = stmt.executeQuery(String.format("select userID from blog where blogID=%d;",blogID));
+	rs.next();
+	String bloguid = Integer.toString(rs.getInt("userID"));
+	String myuid = session.getAttribute("userID").toString();
+	if(!myuid.equals(bloguid)){
+		conn.close();
+		conn = null;
+		throw new Exception("不能删除他人的博客！");
+	}
+	conn.setAutoCommit(false);
 	stmt.executeQuery(String.format("select catID,tagID from blog where blogID=%d into @cid,@tid;",blogID));
 	stmt.executeUpdate(String.format("delete from blog where blogID = %d;", blogID));
 	stmt.executeUpdate(String.format("delete from cat where catID = @cid;"));
